# ui/src/core/api/request/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Database Schema API",
    "version": "v1"
  },
  "paths": {
    "/api/Database/structure": {
      "get": {
        "tags": [
          "Database"
        ],
        "parameters": [
          {
            "name": "connectionName",
            "in": "query",
            "schema": {
              "type": "string",
              "default": "DefaultConnection"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TableInfo"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TableInfo"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TableInfo"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/Database/tableStructure": {
      "get": {
        "tags": [
          "Database"
        ],
        "parameters": [
          {
            "name": "connectionName",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "schema",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "table",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TableInfo"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TableInfo"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TableInfo"
                }
              }
            }
          }
        }
      }
    },
    "/api/Database/GetMergeSql": {
      "post": {
        "tags": [
          "Database"
        ],
        "parameters": [
          {
            "name": "connectionName",
            "in": "query",
            "schema": {
              "type": "string",
              "default": "DefaultConnection"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecordData"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RecordData"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/RecordData"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/Database/tablePaginationRecords": {
      "get": {
        "tags": [
          "Database"
        ],
        "parameters": [
          {
            "name": "connectionName",
            "in": "query",
            "schema": {
              "type": "string",
              "default": "DefaultConnection"
            }
          },
          {
            "name": "schema",
            "in": "query",
            "schema": {
              "type": "string",
              "default": "Config"
            }
          },
          {
            "name": "table",
            "in": "query",
            "schema": {
              "type": "string",
              "default": "Config"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "perPage",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 100
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/StringObjectIDictionaryPagedResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StringObjectIDictionaryPagedResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/StringObjectIDictionaryPagedResult"
                }
              }
            }
          }
        }
      }
    },
    "/api/Database/tableRecord": {
      "get": {
        "tags": [
          "Database"
        ],
        "parameters": [
          {
            "name": "connectionName",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "schema",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "table",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "recordId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "object",
                  "additionalProperties": { }
                }
              },
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": { }
                }
              },
              "text/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": { }
                }
              }
            }
          }
        }
      }
    },
    "/api/Database/connection/{connectionName}": {
      "post": {
        "tags": [
          "Database"
        ],
        "parameters": [
          {
            "name": "connectionName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "string"
              }
            },
            "text/json": {
              "schema": {
                "type": "string"
              }
            },
            "application/*+json": {
              "schema": {
                "type": "string"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Database/connections": {
      "get": {
        "tags": [
          "Database"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ConnectionRecord"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ConnectionRecord"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ConnectionRecord"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/Database/columnConfig/{connectionName}": {
      "put": {
        "tags": [
          "Database"
        ],
        "parameters": [
          {
            "name": "connectionName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/CustomColumnInfo"
                }
              }
            },
            "text/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/CustomColumnInfo"
                }
              }
            },
            "application/*+json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/CustomColumnInfo"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ColumnInfo": {
        "type": "object",
        "properties": {
          "tableId": {
            "type": "integer",
            "format": "int32"
          },
          "schema": {
            "type": "string",
            "nullable": true
          },
          "table": {
            "type": "string",
            "nullable": true
          },
          "columnId": {
            "type": "integer",
            "format": "int32"
          },
          "columnName": {
            "type": "string",
            "nullable": true
          },
          "dataType": {
            "type": "string",
            "nullable": true
          },
          "dbType": {
            "type": "string",
            "nullable": true
          },
          "defaultValue": {
            "type": "string",
            "nullable": true
          },
          "isNullable": {
            "type": "boolean"
          },
          "isIdentity": {
            "type": "boolean"
          },
          "isPK": {
            "type": "boolean"
          },
          "isFK": {
            "type": "boolean"
          },
          "isUK": {
            "type": "boolean"
          },
          "isExtension": {
            "type": "boolean"
          },
          "isCustom": {
            "type": "boolean"
          },
          "schemaFK": {
            "type": "string",
            "nullable": true
          },
          "tableFK": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ConnectionRecord": {
        "type": "object",
        "properties": {
          "connectionName": {
            "type": "string",
            "nullable": true
          },
          "connectionString": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "CustomColumnInfo": {
        "type": "object",
        "properties": {
          "connectionName": {
            "type": "string",
            "nullable": true
          },
          "schema": {
            "type": "string",
            "nullable": true
          },
          "table": {
            "type": "string",
            "nullable": true
          },
          "columnName": {
            "type": "string",
            "nullable": true
          },
          "dataType": {
            "type": "string",
            "nullable": true
          },
          "dbType": {
            "type": "string",
            "nullable": true
          },
          "defaultValue": {
            "type": "string",
            "nullable": true
          },
          "isNullable": {
            "type": "boolean",
            "nullable": true
          },
          "isIdentity": {
            "type": "boolean",
            "nullable": true
          },
          "isPK": {
            "type": "boolean",
            "nullable": true
          },
          "isFK": {
            "type": "boolean",
            "nullable": true
          },
          "isUK": {
            "type": "boolean",
            "nullable": true
          },
          "isExtension": {
            "type": "boolean",
            "nullable": true
          },
          "schemaFK": {
            "type": "string",
            "nullable": true
          },
          "tableFK": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "RecordData": {
        "type": "object",
        "properties": {
          "tableId": {
            "type": "integer",
            "format": "int32"
          },
          "parentColumn": {
            "type": "string",
            "nullable": true
          },
          "columns": {
            "type": "object",
            "additionalProperties": { },
            "nullable": true
          },
          "dependencies": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RecordData"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "StringObjectIDictionaryPagedResult": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "type": "object",
              "additionalProperties": { }
            },
            "nullable": true
          },
          "totalCount": {
            "type": "integer",
            "format": "int32"
          },
          "pageNumber": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32",
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
      "TableInfo": {
        "type": "object",
        "properties": {
          "tableId": {
            "type": "integer",
            "format": "int32"
          },
          "schema": {
            "type": "string",
            "nullable": true
          },
          "table": {
            "type": "string",
            "nullable": true
          },
          "isEntity": {
            "type": "boolean"
          },
          "isExtension": {
            "type": "boolean"
          },
          "identityColumn": {
            "type": "string",
            "nullable": true
          },
          "representationColumns": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "columns": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ColumnInfo"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    }
  }
}